+++ Bot.Business/Models/MuteDto.cs
@@ -8,11 +8,11 @@ namespace Bot.Business.Models
 {
     public class MuteDto
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
-        public int ChatId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
+        public long ChatId { get; set; }
         public int MuteDuration { get; set; }
         public int MuteTime { get; set; }
-        public int MessageId { get; set; }
+        public long MessageId { get; set; }
     }
 }
