+++ Bot.Business/Models/ReportDto.cs
@@ -8,11 +8,11 @@ namespace Bot.Business.Models
 {
     public class ReportDto
     {
-        public int Id { get; set; }
-        public int SourceUserId { get; set; }
-        public int TargetUserId { get; set; }
-        public int WarningMessage { get; set; }
-        public int ReportedMessage { get; set; }
+        public long Id { get; set; }
+        public long SourceUserId { get; set; }
+        public long TargetUserId { get; set; }
+        public long WarningMessage { get; set; }
+        public long ReportedMessage { get; set; }
 
     }
 }
